Add HasAuthenticatedCreator.php trait without global scope
